Undo StatsContoller.class refactor
Undo typo correction in StatsContoller.class to prevent conflicts
